feat(basic): 페이지 컴포넌트 분리
- `ShoppingPage.tsx`: 상품 목록, 장바구니, 쿠폰 적용, 결제 로직 등 쇼핑몰 관련 기능을 담당합니다.
- `AdminPage.tsx`: 상품 및 쿠폰 관리 등 관리자 관련 기능을 담당합니다.
- `App.tsx`는 이제 `isAdmin` 상태에 따라 두 페이지 컴포넌트를 렌더링하는 역할만 수행합니다.
- `addNotification` 함수가 필요한 하위 컴포넌트 및 훅으로 올바르게 전달되도록 수정했습니다.
