+++ src/basic/components/ui/Button.tsx
@@ -1,5 +1,5 @@
 interface ButtonProps {
-  children: React.ReactNode;
+  children?: React.ReactNode;
   onClick?: () => void;
   disabled?: boolean;
   variant?: 'primary' | 'secondary' | 'danger';
